Fixed the handling of bits for the magnetometer readings.  They look plausible now, though I haven't tested them thoroughly.

+++ LSM303.cpp
@@ -101,9 +101,9 @@ void LSM303::readMag(void)
     }
 
     // DLM, DLHC: register address order is X,Z,Y with high bytes first
-    m(0) = block[1] + (block[0] << 8);
-    m(1) = block[5] + (block[4] << 8);
-    m(2) = block[3] + (block[2] << 8);
+    m(0) = (int16_t)(block[1] + (block[0] << 8));
+    m(1) = (int16_t)(block[5] + (block[4] << 8));
+    m(2) = (int16_t)(block[3] + (block[2] << 8));
 }
 
 // Reads all 6 channels of the LSM303 and stores them in the object variables
